refactor(dsg): renaming functions
