+++ packages/amplication-data-service-generator/src/server/resource/dto/create-nested/create-create-nested-many-without-input/index.ts
@@ -1,2 +1,1 @@
 export * from "./create-create-nested-many-without-input";
-export * from "./create-ast-id";
