+++ packages/amplication-data-service-generator/src/server/resource/dto/create-nested/create-nested.ts
@@ -2,15 +2,36 @@ import { builders, namedTypes } from "ast-types";
 import { TSTypeKind } from "ast-types/gen/kinds";
 import { ObjectField, ScalarField } from "prisma-schema-dsl";
 import { Entity, EntityField } from "../../../../types";
-import { classProperty } from "../../../../util/ast";
+import {
+  classDeclaration,
+  classProperty,
+  NamedClassDeclaration,
+} from "../../../../util/ast";
 import { createPrismaFields } from "../../../prisma/create-prisma-schema";
 import { ApiPropertyDecoratorBuilder } from "../api-property-decorator";
 import {
   createFieldValueTypeFromPrismaField,
   getTypeName,
 } from "../create-field-class-property";
-import { createGraphQLFieldDecorator } from "../graphql-field-decorator";
 import { EntityDtoTypeEnum } from "../entity-dto-type-enum";
+import { createGraphQLFieldDecorator } from "../graphql-field-decorator";
+import { INPUT_TYPE_DECORATOR } from "../nestjs-graphql.util";
+
+export function createNestedInputDTO(
+  classId: namedTypes.Identifier,
+  entity: Entity,
+  toManyField: EntityField,
+  dtoType: EntityDtoTypeEnum
+): NamedClassDeclaration {
+  const properties = createNestedManyProperties(toManyField, entity, dtoType);
+  const decorators = properties.length ? [INPUT_TYPE_DECORATOR] : [];
+  return classDeclaration(
+    classId,
+    builders.classBody(properties),
+    null,
+    decorators
+  ) as NamedClassDeclaration;
+}
 
 export enum NestedMutationOptions {
   "Create" = "create",
@@ -18,7 +39,7 @@ export enum NestedMutationOptions {
   "ConnectOrCreate" = "connectOrCreate",
 }
 
-export function createNestedManyProperties(
+function createNestedManyProperties(
   field: EntityField,
   entity: Entity,
   dtoType: EntityDtoTypeEnum
