+++ packages/amplication-data-service-generator/src/server/resource/dto/create-nested/create-update-many-without-input/index.ts
@@ -1,2 +1,1 @@
 export * from "./create-update-many-without-input";
-export * from "./create-ast-id";
