--- packages/amplication-data-service-generator/src/server/resource/dto/to-many/create-create-nested-many-without-input/create-ast-id.ts
@@ -1,12 +1,0 @@
-import { builders, namedTypes } from "ast-types";
-import { pascalCase } from "pascal-case";
-export function createCreateNestedManyWithoutInputID(
-  pluralEntityName: string,
-  nestedEntityName: string
-): namedTypes.Identifier {
-  return builders.identifier(
-    `${pascalCase(nestedEntityName)}CreateNestedManyWithout${pascalCase(
-      pluralEntityName
-    )}Input`
-  );
-}
